moved WCS-specific functions out of FrameStore and into a utility file. Initial support for zooming on secondary images
