+++ src/components/ImageView/ImageViewComponent.tsx
@@ -121,9 +121,18 @@ export class ImageViewComponent extends React.Component<WidgetProps> {
         const appStore = this.props.appStore;
         if (appStore.activeFrame) {
             const zoomSpeed = 1 + Math.abs(delta / 750.0);
-            const newZoom = appStore.activeFrame.zoomLevel * (delta > 0 ? zoomSpeed : 1.0 / zoomSpeed);
-            // Shift from one-indexed image space position to zero-indexed
-            appStore.activeFrame.zoomToPoint(cursorInfo.posImageSpace.x + 1, cursorInfo.posImageSpace.y + 1, newZoom);
+
+            // If frame is spatially matched, apply zoom to the reference frame, rather than the active frame
+            if (appStore.activeFrame.spatialReference) {
+                const newZoom = appStore.activeFrame.spatialReference.zoomLevel * (delta > 0 ? zoomSpeed : 1.0 / zoomSpeed);
+                // Shift from one-indexed image space position to zero-indexed
+                // TODO: apply zoom point properly based on rotated, scaled and translated cursor location
+                appStore.activeFrame.spatialReference.zoomToPoint(cursorInfo.posImageSpace.x + 1, cursorInfo.posImageSpace.y + 1, newZoom);
+            } else {
+                const newZoom = appStore.activeFrame.zoomLevel * (delta > 0 ? zoomSpeed : 1.0 / zoomSpeed);
+                // Shift from one-indexed image space position to zero-indexed
+                appStore.activeFrame.zoomToPoint(cursorInfo.posImageSpace.x + 1, cursorInfo.posImageSpace.y + 1, newZoom);
+            }
         }
     };
