+++ src/components/ImageView/Toolbar/ToolbarComponent.tsx
@@ -82,7 +82,8 @@ export class ToolbarComponent extends React.Component<ToolbarComponentProps> {
             className += " docked";
         }
 
-        const currentZoomSpan = <span><br/><i><small>Current: {toFixed(frame.zoomLevel, 2)}x</small></i></span>;
+        const zoomLevel = (frame.spatialReference && frame.spatialTransform) ? frame.spatialReference.zoomLevel * frame.spatialTransform.scale.x : frame.zoomLevel;
+        const currentZoomSpan = <span><br/><i><small>Current: {toFixed(zoomLevel, 2)}x</small></i></span>;
         const tooltipPosition: PopoverPosition = this.props.vertical ? "left" : "bottom";
 
         const regionMenu = (
@@ -156,7 +157,7 @@ export class ToolbarComponent extends React.Component<ToolbarComponentProps> {
                     <Button icon="zoom-to-fit" onClick={frame.fitZoom}/>
                 </Tooltip>
                 <Tooltip position={tooltipPosition} content={<span>Debug: Toggle WCS ref</span>}>
-                    <AnchorButton icon="link" active={!!frame.spatialReference} disabled={frame === appStore.spatialReference} onClick={this.handleSpatialReferenceToggled}/>
+                    <Button icon="link" active={!!frame.spatialReference} disabled={frame === appStore.spatialReference} onClick={this.handleSpatialReferenceToggled}/>
                 </Tooltip>
                 <Tooltip position={tooltipPosition} content={<span>Overlay Coordinate <br/><small><i>Current: {ToolbarComponent.CoordinateSystemTooltip.get(coordinateSystem)}</i></small></span>}>
                     <Popover content={coordinateSystemMenu} position={Position.TOP} minimal={true}>
