+++ src/models/Transform2D.ts
@@ -1,0 +1,7 @@
+import {Point2D} from "./Point2D";
+
+export interface Transform2D {
+    translation: Point2D;
+    rotation: number;
+    scale: Point2D;
+}
