+++ src/models/index.ts
@@ -3,6 +3,7 @@ export * from "./ChannelInfo";
 export * from "./ChannelType";
 export * from "./FrameView";
 export * from "./Point2D";
+export * from "./Transform2D";
 export * from "./SpectralInfo";
 export * from "./TileCoordinate";
 export * from "./Theme";
