+++ src/utilities/index.ts
@@ -6,3 +6,4 @@ export * from "./parsing";
 export * from "./tiling";
 export * from "./units";
 export * from "./webgl";
+export * from "./wcs";
